+++ Cindy3D/src/de/tum/in/cindy3dplugin/jogl/primitives/Circle.java
@@ -11,7 +11,7 @@ public class Circle extends Primitive {
 	
 	public Circle(double centerX, double centerY, double centerZ,
 			double normalX, double normalY, double normalZ, double radius,
-			Color color, int shininess, double alpha) {
+			Color color, double shininess, double alpha) {
 		super(color, shininess, alpha);
 		
 		center = new Vector3D(centerX, centerY, centerZ);
